+++ central/cluster/datastore/mocks/datastore.go
@@ -12,7 +12,6 @@ import (
 	concurrency "github.com/stackrox/rox/pkg/concurrency"
 	search "github.com/stackrox/rox/pkg/search"
 	reflect "reflect"
-	time "time"
 )
 
 // MockDataStore is a mock of DataStore interface
@@ -215,25 +214,6 @@ func (m *MockDataStore) UpdateClusterCertExpiryStatus(arg0 context.Context, arg1
 func (mr *MockDataStoreMockRecorder) UpdateClusterCertExpiryStatus(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpdateClusterCertExpiryStatus", reflect.TypeOf((*MockDataStore)(nil).UpdateClusterCertExpiryStatus), arg0, arg1, arg2)
-}
-
-// UpdateClusterContactTimes mocks base method
-func (m *MockDataStore) UpdateClusterContactTimes(arg0 context.Context, arg1 time.Time, arg2 ...string) error {
-	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1}
-	for _, a := range arg2 {
-		varargs = append(varargs, a)
-	}
-	ret := m.ctrl.Call(m, "UpdateClusterContactTimes", varargs...)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// UpdateClusterContactTimes indicates an expected call of UpdateClusterContactTimes
-func (mr *MockDataStoreMockRecorder) UpdateClusterContactTimes(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1}, arg2...)
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpdateClusterContactTimes", reflect.TypeOf((*MockDataStore)(nil).UpdateClusterContactTimes), varargs...)
 }
 
 // UpdateClusterHealth mocks base method